requirements.txt
vllm
openai
google-generativeai
scikit-learn
numpy
sentence-transformers